+++ src/Bzn.Cloudios.Application/Abstractions/ITenantProvider.cs
@@ -1,0 +1,8 @@
+namespace Bzn.Cloudios.Application.Abstractions;
+
+public interface ITenantProvider
+{
+    Guid RealmId { get; }
+    string Role { get; }
+    Guid UserId { get; }
+}